+++ xenon-common/src/main/java/com/vmware/xenon/common/ServiceHost.java
@@ -51,6 +51,7 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ForkJoinPool;
 import java.util.concurrent.ForkJoinWorkerThread;
+import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
@@ -65,6 +66,14 @@ import java.util.logging.Logger;
 
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.TrustManagerFactory;
+
+import io.opentracing.ActiveSpan;
+import io.opentracing.SpanContext;
+import io.opentracing.Tracer;
+import io.opentracing.propagation.Format;
+import io.opentracing.propagation.TextMapExtractAdapter;
+import io.opentracing.propagation.TextMapInjectAdapter;
+import io.opentracing.tag.Tags;
 
 import com.vmware.xenon.common.FileUtils.ResourceEntry;
 import com.vmware.xenon.common.NodeSelectorService.SelectAndForwardRequest;
@@ -90,6 +99,8 @@ import com.vmware.xenon.common.http.netty.NettyHttpServiceClient;
 import com.vmware.xenon.common.jwt.JWTUtils;
 import com.vmware.xenon.common.jwt.Signer;
 import com.vmware.xenon.common.jwt.Verifier;
+import com.vmware.xenon.common.opentracing.TracerFactory;
+import com.vmware.xenon.common.opentracing.TracingUtils;
 import com.vmware.xenon.services.common.AuthCredentialsService;
 import com.vmware.xenon.services.common.AuthorizationContextService;
 import com.vmware.xenon.services.common.AuthorizationTokenCacheService;
@@ -585,6 +596,12 @@ public class ServiceHost implements ServiceRequestSender {
     private final Set<String> pendingServiceDeletions = Collections
             .synchronizedSet(new HashSet<String>());
 
+    /**
+     * OpenTracing tracer. Currently spans have recount semantics, unlike OperationContext, so rather than
+     * being more restrictive than OperationContext, we just track such spans independently.
+     */
+    private final Tracer otTracer;
+
     private OperationProcessingChain opProcessingChain;
     private AuthorizationFilter authorizationFilter;
 
@@ -635,6 +652,7 @@ public class ServiceHost implements ServiceRequestSender {
     protected ServiceHost() {
         this.state = new ServiceHostState();
         this.state.id = UUID.randomUUID().toString();
+        this.otTracer = TracerFactory.factory.create(this);
     }
 
     public ServiceHost initialize(String[] args) throws Throwable {
@@ -1402,8 +1420,11 @@ public class ServiceHost implements ServiceRequestSender {
         return Executors.newFixedThreadPool(threadCount, new NamedThreadFactory(s.getUri().toString()));
     }
 
+    @SuppressWarnings("try")
     public ServiceHost start() throws Throwable {
-        return startImpl();
+        try (ActiveSpan span = this.otTracer.buildSpan("ServiceHost.start").startActive()) {
+            return startImpl();
+        }
     }
 
     private void setSystemProperties() {
@@ -3515,6 +3536,7 @@ public class ServiceHost implements ServiceRequestSender {
     /**
      * Infrastructure use only
      */
+    @SuppressWarnings("try")
     public boolean handleRequest(Service service, Operation inboundOp) {
         if (inboundOp == null && service != null) {
             inboundOp = service.dequeueRequest();
@@ -3540,13 +3562,40 @@ public class ServiceHost implements ServiceRequestSender {
             return true;
         }
 
-        OperationProcessingContext context = this.opProcessingChain.createContext(this);
-        final Operation finalInboundOp = inboundOp;
-        this.opProcessingChain.processRequest(inboundOp, context, o -> {
-            handleRequestAfterOpProcessingChain(context.getService(), finalInboundOp);
-        });
+        /* Create a tracing span for this new request we're handling */
+        SpanContext extractedContext = this.otTracer.extract(
+                Format.Builtin.HTTP_HEADERS, new TextMapExtractAdapter(inboundOp.getRequestHeaders()));
+        try (ActiveSpan span = this.otTracer.buildSpan(inboundOp.getUri().getPath().toString())
+                // By definition this is a new request, so we don't want to use any active span (e.g. due to
+                // fastpathing or a bug or some such) as a parent.
+                .ignoreActiveSpan()
+                .asChildOf(extractedContext)
+                .withTag(Tags.SPAN_KIND.getKey(), Tags.SPAN_KIND_SERVER)
+                .startActive()) {
+            // Populate common operation tags for the span.
+            TracingUtils.setSpanTags(inboundOp, span);
 
-        return true;
+            // capture the response code of the operation
+            final ActiveSpan.Continuation completion_cont = span.capture();
+            inboundOp.nestCompletionCloneSafe((o, e) -> {
+                try (ActiveSpan after_span = completion_cont.activate()) {
+                    after_span.setTag(Tags.HTTP_STATUS.getKey(), Integer.toString(o.getStatusCode()));
+                    if (e == null) {
+                        o.complete();
+                    } else {
+                        o.fail(e);
+                    }
+                }
+            });
+
+            // Pass the operation through the processing chain.
+            OperationProcessingContext context = this.opProcessingChain.createContext(this);
+            final Operation finalInboundOp = inboundOp;
+            this.opProcessingChain.processRequest(inboundOp, context, o -> {
+                handleRequestAfterOpProcessingChain(context.getService(), finalInboundOp);
+            });
+            return true;
+        }
     }
 
     private void handleRequestAfterOpProcessingChain(Service service, Operation inboundOp) {
@@ -3615,17 +3664,33 @@ public class ServiceHost implements ServiceRequestSender {
 
     private void queueOrScheduleRequestInternal(Service s, Operation op) {
         if (!s.queueRequest(op)) {
+            ActiveSpan span = this.otTracer.activeSpan();
+            ActiveSpan.Continuation cont = null != span ? span.capture() : null;
             Runnable r = () -> {
                 OperationContext opCtx = extractAndApplyContext(op);
                 try {
-                    s.handleRequest(op);
+                    ActiveSpan contspan = null != cont ? cont.activate() : null;
+                    try {
+                        s.handleRequest(op);
+                    } finally {
+                        if (contspan != null) {
+                            contspan.close();
+                        }
+                    }
                 } catch (Exception e) {
                     handleUncaughtException(s, op, e);
                 } finally {
                     OperationContext.setFrom(opCtx);
                 }
             };
-            this.executor.execute(r);
+            try {
+                this.executor.execute(r);
+            } catch (RejectedExecutionException e) {
+                if (cont != null) {
+                    cont.activate().close();
+                }
+                throw e;
+            }
         }
     }
 
@@ -3655,7 +3720,30 @@ public class ServiceHost implements ServiceRequestSender {
             return;
         }
 
-        c.send(op);
+        // Trace the request we're about to send.
+        // We don't have enough information here to set a great operationName; in future we'll want to provide a way
+        // to customise it e.g. via properties on the operation.
+        try (ActiveSpan span = this.otTracer.buildSpan(op.getUri().getPath().toString())
+                .withTag(Tags.SPAN_KIND.getKey(), Tags.SPAN_KIND_CLIENT)
+                .startActive()) {
+            TracingUtils.setSpanTags(op, span);
+            // Pass the span into the network request, propagating it across hosts.
+            this.otTracer.inject(span.context(), Format.Builtin.HTTP_HEADERS,
+                    new TextMapInjectAdapter(op.getRequestHeaders()));
+            final ActiveSpan.Continuation cont = span.capture();
+            // Capture the HTTP status code into the span.
+            op.nestCompletionCloneSafe((o, e) -> {
+                try (ActiveSpan contspan = cont.activate()) {
+                    contspan.setTag(Tags.HTTP_STATUS.getKey(), Integer.toString(o.getStatusCode()));
+                    if (e == null) {
+                        o.complete();
+                    } else {
+                        o.fail(e);
+                    }
+                }
+            });
+            c.send(op);
+        }
     }
 
     private void traceOperation(Operation op) {
@@ -4515,10 +4603,27 @@ public class ServiceHost implements ServiceRequestSender {
             throw new IllegalStateException("Stopped");
         }
         OperationContext origContext = OperationContext.getOperationContext();
-        executor.execute(() -> {
-            OperationContext.setFrom(origContext);
-            executeRunnableSafe(task);
-        });
+        /* Pass any current tracing span potentially cross-thread */
+        ActiveSpan span = this.otTracer.activeSpan();
+        final ActiveSpan.Continuation cont = null != span ? span.capture() : null;
+        try {
+            executor.execute(() -> {
+                ActiveSpan contspan = null != cont ? cont.activate() : null;
+                try {
+                    OperationContext.setFrom(origContext);
+                    executeRunnableSafe(task);
+                } finally {
+                    if (contspan != null) {
+                        contspan.close();
+                    }
+                }
+            });
+        } catch (RejectedExecutionException e) {
+            if (cont != null) {
+                cont.activate().close();
+            }
+            throw e;
+        }
     }
 
     /**
@@ -5451,4 +5556,20 @@ public class ServiceHost implements ServiceRequestSender {
                 ServiceErrorResponse.ERROR_CODE_SERVICE_ALREADY_EXISTS,
                 e);
     }
+
+
+    /**
+     * Get the {@link io.opentracing.Tracer} tracer this host is using.
+     * <p>
+     *     This can be used to get the active span source via getTracer.activeSpan(), which
+     *     allows service developers to create spans around any thing that they want to instrument - whether
+     *     thats aggregations of regular Operations, or other tasks like custom IO handling, library callouts, or
+     *     for use with their own executor hand-offs.
+     * </p>
+     * @return
+     */
+    public Tracer getTracer() {
+        return this.otTracer;
+    }
+
 }